move AccountInfo back to shared
