+++ fendermint/actors/blobs/src/lib.rs
@@ -8,4 +8,3 @@ mod sol_facade;
 mod state;
 
 pub use shared::*;
-pub use state::*;
