+++ fendermint/actors/blobs/src/state.rs
@@ -2,7 +2,7 @@
 // Copyright 2021-2023 Protocol Labs
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::error::Error;
 use std::fmt::Display;
 use std::str::from_utf8;
@@ -34,8 +34,6 @@ mod expiries;
 use accounts::AccountsState;
 use blobs::{BlobsProgressCollection, BlobsState};
 use expiries::{ExpiriesState, ExpiryUpdate};
-use fil_actors_runtime::runtime::Runtime;
-use recall_actor_sdk::to_delegated_address;
 
 /// The state represents all accounts and stored blobs.
 #[derive(Debug, Serialize_tuple, Deserialize_tuple)]
@@ -1799,66 +1797,6 @@ fn ensure_gas_limit(
         }
     }
     Ok(())
-}
-
-/// The return type used for Account.
-#[derive(Debug, Serialize_tuple, Deserialize_tuple)]
-pub struct AccountInfo {
-    /// Total size of all blobs managed by the account.
-    pub capacity_used: u64,
-    /// Current free credit in byte-blocks that can be used for new commitments.
-    pub credit_free: Credit,
-    /// Current committed credit in byte-blocks that will be used for debits.
-    pub credit_committed: Credit,
-    /// Optional default sponsor account address.
-    pub credit_sponsor: Option<Address>,
-    /// The chain epoch of the last debit.
-    pub last_debit_epoch: ChainEpoch,
-    /// Credit approvals to other accounts from this account, keyed by receiver.
-    pub approvals_to: HashMap<Address, CreditApproval>,
-    /// Credit approvals to this account from other accounts, keyed by sender.
-    pub approvals_from: HashMap<Address, CreditApproval>,
-    /// The maximum allowed TTL for actor's blobs.
-    pub max_ttl: ChainEpoch,
-    /// The total token value an account has used to buy credits.
-    pub gas_allowance: TokenAmount,
-}
-
-impl AccountInfo {
-    pub fn from(rt: &impl Runtime, account: Account) -> Result<Self, ActorError> {
-        let store = rt.store();
-        let mut approvals_to = HashMap::new();
-        account
-            .approvals_to
-            .hamt(store)?
-            .for_each(|address, approval| {
-                let external_account_address = to_delegated_address(rt, address)?;
-                approvals_to.insert(external_account_address, approval.clone());
-                Ok(())
-            })?;
-
-        let mut approvals_from = HashMap::new();
-        account
-            .approvals_from
-            .hamt(store)?
-            .for_each(|address, approval| {
-                let external_account_address = to_delegated_address(rt, address)?;
-                approvals_from.insert(external_account_address, approval.clone());
-                Ok(())
-            })?;
-
-        Ok(AccountInfo {
-            capacity_used: account.capacity_used,
-            credit_free: account.credit_free,
-            credit_committed: account.credit_committed,
-            credit_sponsor: account.credit_sponsor,
-            last_debit_epoch: account.last_debit_epoch,
-            approvals_to,
-            approvals_from,
-            max_ttl: account.max_ttl,
-            gas_allowance: account.gas_allowance,
-        })
-    }
 }
 
 #[cfg(test)]
